Kotlinify Chucker sample app
